changed discrete uniform distribution for non-discrete exponential

+++ include/average_stats.hpp
@@ -14,29 +14,29 @@ class AverageStats {
         void draw_delay_plot();
         void draw_scheduling_plot();
 
-        float average_total_time = 0;
-        int common_total_time = 0;
+        double average_total_time = 0;
+        double common_total_time = 0;
 
-        float average_total_idle_time = 0;
-        int common_total_idle_time = 0;
+        double average_total_idle_time = 0;
+        double common_total_idle_time = 0;
 
-        float average_total_skip_time = 0;
-        int common_total_skip_time = 0;
+        double average_total_skip_time = 0;
+        double common_total_skip_time = 0;
 
-        float average_total_processing_time = 0;
-        int common_total_processing_time = 0;
+        double average_total_processing_time = 0;
+        double common_total_processing_time = 0;
 
-        float average_total_packet_count = 0;
-        int common_total_packet_count = 0;
+        double average_total_packet_count = 0;
+        double common_total_packet_count = 0;
 
-        float average_total_lost_packet_count = 0;
-        int common_total_lost_packet_count = 0;
+        double average_total_lost_packet_count = 0;
+        double common_total_lost_packet_count = 0;
     
-        float average_total_retried_packet_count = 0;
-        int common_total_retried_packet_count = 0;
+        double average_total_retried_packet_count = 0;
+        double common_total_retried_packet_count = 0;
 
-        std::vector<float> average_delays;
-        std::map<int, std::vector<float>> average_packets_scheduled_by_ms;
+        std::vector<double> average_delays;
+        std::map<double, std::vector<double>> average_packets_scheduled_by_ms;
 
         std::vector<ExecutionStats> stats_array;
